requirements.txt
beautifulsoup4>=4.9.3,<5.0
Flask>=2.3.0
Jinja2>=3.1.2
Werkzeug>=2.3.0
googlesearch-python>=1.0.1
numpy==1.26.4
pandas==1.5.3
python_dateutil==2.8.2
requests==2.28.2
scikit-learn==1.2.2
whois==0.9.13
gunicorn==20.1.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
